multi-agent-chatbot-lollypop-design-v3/requirements.txt
python-dotenv==1.2.1
ipython==9.7.0
PyMuPDF==1.26.6
sentence-transformers==5.1.2
python-docx==1.2.0
faiss-cpu==1.13.0
langgraph-checkpoint-sqlite==2.0.11
tiktoken==0.12.0
flask-cors==6.0.1
Flask==3.1.2
gunicorn==23.0.0
email-validator==2.3.0
dnspython==2.8.0
Werkzeug==3.1.3
APScheduler==3.11.1
pandas==2.3.3
bleach==6.3.0
google-cloud-storage==3.6.0
msgpack==1.1.2
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-experimental==0.3.4
langchain-google-community[gmail]==2.0.7
langchain-milvus==0.2.1
langchain-openai==0.3.32
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.10
langdetect==1.0.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.24
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.20
boto3==1.35.95
pyyaml==6.0.1
beautifulsoup4==4.12.3
